refactor: refactor regex paring to lex then parse
Refactor the regex parser to first convert a regular expression into a
list of tokens using a lexer. The list of tokens is then parsed using a
seperate parser to an internal representation of a regular expression.
This improves the seperation between the string representation and the
NFA that will be built from the regex.